Rename compiler/compiler.ts to compiler/index.ts
