+++ src/compiler/index.ts
